refactor: Move dates to separate module
Circular imports, apparently.
